replaced autoboxing statements for 1.4
git-svn-id: https://svn.broadinstitute.org/genepattern/trunk@3015 48a7a3f7-ad46-0410-b969-29950918272e

+++ src/org/genepattern/data/expr/MetaData.java
@@ -144,10 +144,11 @@ public class MetaData {
      */
     protected int getDepth(String name) {
         Integer depth = (Integer) metaDataName2Depth.get(name);
+        
         if (depth == null) {
             depth = new Integer(metaData.size());
-            metaData.ensureCapacity(depth);
-            metaData.set(depth, new String[size]);
+            metaData.ensureCapacity(depth.intValue());
+            metaData.set(depth.intValue(), new String[size]);
             metaDataName2Depth.put(name, depth);
         }
         return depth.intValue();
